requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
https://gradio-builds.s3.amazonaws.com/bed454c3d22cfacedc047eb3b0ba987b485ac3fd/gradio-4.40.0-py3-none-any.whl
torch==2.1.2 
torchvision==0.16.2
torchaudio==2.1.2
diffusers==0.27.2
accelerate==0.28.0
tensorflow==2.14.0
tensorboard==2.14.0
opencv-python==4.9.0.80
soundfile==0.12.1
transformers==4.41.1
gdown==5.2.0
requests==2.32.3
imageio==2.35.1
imageio[ffmpeg]
omegaconf==2.3.0
ffmpeg-python==0.2.0
spaces==0.30.0
moviepy==1.0.3
numpy==1.23.5
scipy==1.13.1
librosa==0.9.2
numba==0.56.4
pytorch-lightning==2.4.0
onnxruntime==1.19.2; sys_platform == 'darwin'
onnxruntime-gpu==1.19.2; sys_platform != 'darwin'
tqdm==4.66.5
funasr==1.1.6
cn2an==0.5.22
pypinyin==0.52.0
pyopenjtalk==0.3.4
g2p-en==2.1.0
modelscope==1.10.0
sentencepiece==0.2.0
chardet==5.2.0
PyYAML==6.0.2
psutil==5.9.8
jieba_fast==0.53
jieba==0.42.1
LangSegment==0.3.5
Faster_Whisper==1.0.3
wordsegment==1.3.1
rotary-embedding-torch==0.7.0
pyjyutping==1.0.0
g2pk2==0.0.3
ko-pron==1.3
opencc; sys_platform != 'linux'
opencc==1.1.1; sys_platform == 'linux'
python_mecab_ko==1.3.7; sys_platform != 'win32'
openmim==0.3.9
openai==1.43.0
fastapi[all]
nltk
modelscope_studio
pydub
dashscope
loguru